refactor : #388 <Common> AOP 기능 구현
- 권한을 체크하기 위해 Role 열거형 인스턴스를 가져왔습니다.
